// src/types/network.ts
/**
 * Values are POA Bridge specific
 */
export enum BlockchainEnum {
  NEAR = "near:mainnet",

  ETHEREUM = "eth:1",
  BASE = "eth:8453",
  ARBITRUM = "eth:42161",
  TURBOCHAIN = "eth:1313161567",
  AURORA = "eth:1313161554",
  GNOSIS = "eth:100",
  BERACHAIN = "eth:80094",

  SOLANA = "sol:mainnet",

  BITCOIN = "btc:mainnet",

  DOGECOIN = "doge:mainnet",

  XRPLEDGER = "xrp:mainnet",

  ZCASH = "zec:mainnet",

  TRON = "tron:mainnet",

  SUI = "sui:mainnet",
  OP = "op:mainnet",
  AVAX = "avax:mainnet",
  CARDANO = "cardano:mainnet",
}